services/document-analysis-service/requirements.txt
flask==3.0.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0
nltk==3.8.1
textstat==0.7.3